Fix issue where null values aren't matched with nullable parameters
If `create()` or `constructParams` pass `null` to a constructor
parameter that is nullable, type-hinted, and scalar, the `null` value is
ignored, and if there is a subsequent non-null argument of the hinted
type, `getParams()` uses it instead, discarding the preceding arguments.

At best, this results in exceptions like "Argument 2 passed to
__construct() must be of the type array or null, bool given" or "Too few
arguments to function". At worst, objects are instantiated with
arguments silently shifted to unintended parameters.

Resolved by matching type-hinted scalar parameters with `null` arguments
when `$param->allowsNull()` is true.