Fix issue where closures are bound to superseded Dice instances
When a setter is called after a closure has been cached, the closure
remains bound to the discarded predecessor of the active `Dice`
instance and continues to use it to share instances and resolve
dependencies. Aside from leaking a little memory, this is not usually a
problem because closures and shared instances are discarded whenever
their respective rules are changed, but in scenarios where rules change
frequently, undesirable outcomes are more likely. Potential impacts
range from subtle to disastrous, depending on the application.

Fix:
- Add a `$dice` argument to all closures and their invocations
- Use `$dice` instead of `$this` in all closures
- Make all closures `static`
- Pass `$this` to cached closures