Fix possible race condition in Dice::create()
If a `Dice` instance is modified while instantiating a class, it is
superseded (and discarded, in most cases) before it returns a result.
This isn't a problem unless the class in question, or another
instantiated subsequently, is a shared instance, because it is not added
to the active container's `$instances` property and is therefore not be
reused.

Fix:
- Add `callMutable()` and use it in setters to modify `clone $this` or
  `$this` as needed
- Only clone `$this` on the first call to `callMutable()` (if at all)
- Apply further changes to the same clone until the first
  `callMutable()` returns